ci: isolate marshal files
